Update AuthService.js
Having this outside of the else if triggers socket authentication on each request. This is due to the interceptor fired on every request re-authenticates the socket. This should only be done if the token is refreshed, or they re-login (the re-login triggers the AUTHENTICATED event, and therefore only the refresh needs it.